pyproject.toml
[project]
name = "virtual-assistant-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.9.0",
    "aiohttp>=3.12.14",
    "chromadb",
    "diffusers>=0.34.0",
    "fastapi>=0.116.1",
    "gradio-client>=1.11.0",
    "hf-xet>=1.1.5",
    "ipykernel>=6.29.5",
    "langchain-chroma>=0.2.5",
    "langchain-community",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.6",
    "langchain-huggingface",
    "langdetect>=1.0.9",
    "langgraph>=0.6.2",
    "openai>=1.97.0",
    "pillow>=11.3.0",
    "python-dotenv",
    "rapidfuzz>=3.13.0",
    "regex>=2024.11.6",
    "replicate>=1.0.7",
    "requests>=2.32.4",
    "safetensors>=0.5.3",
    "scrapegraphai",
    "scrapy>=2.13.3",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "transformers>=4.53.2",
    "uvicorn>=0.35.0",
]
